requirements.txt
Flask>=2.2.3
mysql-connector-python
Flask-WTF
bootstrap-flask